# pkg/api/generated/spec/openapi.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:1323"
info:
  description: Stamp API
  version: 1.0.0
  title: Stamp API
  contact:
    name: whkelvin
tags:
  - name: post
  - name: auth
paths:
  /post:
    post:
      operationId: CreatePost
      tags:
        - post
      summary: create a new post
      description: create a new post
      requestBody:
        description: Create a new post
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostPostRequest"
      responses:
        "201":
          description: post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid Input
      security:
        - jwt: []

  /posts:
    get:
      operationId: GetRecentPosts
      tags:
        - post
      summary: get a list of most recent posts
      description: get a list of most recent posts
      parameters:
        - name: size
          in: query
          required: false
          description: Number of results are included in a page
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: lastFetchedItemId
          in: query
          required: false
          description: Guid of the last fetched item, used to retrive next batch of postings. Empty will return the first page.
          schema:
            type: string
      responses:
        "200":
          description: successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResultSet"
        "400":
          description: Invalid Input
      security: []

  /login:
    post:
      operationId: LogIn
      tags:
        - auth
      summary: log in or sign up to stamp with an external auth provider
      description: log in or sign up to stamp with an external auth provider, returns a jwt for accessing stamp api.
      requestBody:
        description: request body for signing up and logging in. Specify one of the currently supported authProviders ['github'] and accessToken for the auth provider
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogInRequest"
      responses:
        "200":
          description: login sucessful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogInResponse"
        "400":
          description: Invalid Input
      security: []

  /refresh-token:
    post:
      operationId: RefreshToken
      tags:
        - auth
      summary: exchange token in header for a new one
      description: exchange token in header for a new one
      responses:
        "200":
          description: Refresh Sucessful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "400":
          description: Bad Request
      security:
        - jwt: []

externalDocs:
  description: Find out more about Open API specification
  url: "https://swagger.io/specification/"

components:
  schemas:
    PostPostRequest:
      type: object
      required:
        - title
        - description
        - link
        - rootDomain
      properties:
        title:
          type: string
        description:
          type: string
        link:
          type: string
        rootDomain:
          type: string
    Post:
      type: object
      required:
        - id
        - title
        - description
        - link
        - createdDate
        - rootDomain
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        link:
          type: string
        createdDate:
          type: string
        rootDomain:
          type: string
    PostResultSet:
      type: object
      required:
        - count
        - posts
        - pageSize
      properties:
        count:
          type: integer
          format: int32
          x-faker:
            datatype.number:
              min: 1
              max: 100
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        pageSize:
          type: integer
          format: int32
    LogInRequest:
      type: object
      required:
        - authProvider
        - accessToken
      properties:
        authProvider:
          type: string
        accessToken:
          type: string
    LogInResponse:
      type: object
      required:
        - jwt
      properties:
        jwt:
          type: string
    RefreshTokenResponse:
      type: object
      required:
        - jwt
      properties:
        jwt:
          type: string
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT